requirements.txt
flask
python-dotenv
openai
requests
aiohttp